+++ src/verifiers/straightverifier.cpp
@@ -40,17 +40,5 @@ void StraightVerifier::verifyHand(Hand &hand) {
  *  @returns 1 if the player won, 2 if the house won else 0 if it was a tie.
  */
 int StraightVerifier::breakTie(Player player, House house) {
-    // same code as in StraightFlushVerifier::breakTie
-    auto acumCards = [&](const Hand& hand) -> int {
-        const std::vector<Card>& cards = hand.getTopFiveCards();
-        int sum = std::accumulate(cards.cbegin(), cards.cend(), 0, [](int accumulator, const Card& card){
-            return accumulator + card.value;
-        });
-        return sum;
-    };
-    int playerSum = acumCards(player.hand), houseSum = acumCards(house.hand);
-
-    if (playerSum > houseSum) return 1;
-    else if (houseSum > playerSum) return 2;
-    return 0;
+    return breakStraightTypes(player, house);
 }
